[TASK-9] reorganization
