+++ dmitriy.krutikov/task-9/internal/contact/types.go
